allow frac steps and drop latencies if slow in ctlr

+++ src/java/org/apache/cassandra/config/Config.java
@@ -38,7 +38,7 @@ public class Config
     public static final String PROPERTY_PREFIX = "cassandra.";
 
     public String compaction_controller_kind = "dummy";
-    public int compaction_controller_step_size_mbps = 1;
+    public double compaction_controller_step_size_mbps = 1;
     public double compaction_controller_remain_frac = 0.7;
     public double compaction_controller_slo_ms = 0;
     public int compaction_controller_inc_fails_thresh = 0;
